Log water mark for incoming queue in cacher
